fix development points round (66.5 is now 67)

+++ rmoffice/src/main/java/net/sf/rmoffice/core/RMSheet.java
@@ -732,7 +732,7 @@ public class RMSheet extends AbstractPropertyChangeSupport {
 				count++;
 			}
 		}
-		return Math.round( ap / count );
+		return Math.round( (float)ap / count );
 	}
 
 	public int getSkillcategoryStatBonus(SkillCategory category) {
